added routing and fixed dexie DatabaseClosedError
