+++ src/App.svelte
@@ -1,32 +1,17 @@
 <script>
-	import { onMount } from 'svelte';
+	import { Router, Route } from 'svelte-routing';
 
-	export let name;
-	import './tachyons.css';
-	import ItemList from './components/ItemList.svelte';
+	import Home from './components/Home.svelte';
+	import AddEditForm from './components/AddEditForm.svelte';
+	import Navbar from './components/Navbar.svelte';
 
-	import db from './db';
-
-	let applications = [];
-	const tableName = 'applications';
-
-	onMount(async () => {
-		try {
-			// TODO: dexie insists to use the yield keyword instead of await but Svelte is throwing an error
-		    applications = await db.table(tableName).toArray();
-		} finally {
-		    db.close();
-		}
-	});
 </script>
 
-<main>
-	<section class="pv6-ns">
-		<div class="f3 list pl0 mt0 measure-wide-ns center">
-			<ItemList items={applications} />
-		</div>
-	</section>
-</main>
+<Router>
+	<Navbar />
+	<Route path="/edit" component={AddEditForm} />
+	<Route path="/"><Home /></Route>
+</Router>
 
 <style>
 </style>
