+++ src/components/AddEditForm.svelte
@@ -1,0 +1,15 @@
+<script>
+	import { onMount } from 'svelte';
+    export let location;
+	onMount(async () => {
+	});
+
+    console.log('location', location);
+</script>
+
+<main>
+I am the add/edit form!
+</main>
+
+<style>
+</style>
