+++ src/components/Navbar.svelte
@@ -1,0 +1,8 @@
+<script>
+    import {Link} from 'svelte-routing';
+</script>
+
+<nav>
+    <Link to="/">Home</Link>
+    <Link to="edit">Edit</Link>
+</nav>
